Fixed a number of warnings.

+++ Core/External/Compression/Compression.h
@@ -11,12 +11,12 @@ namespace Illumina
 		public:
 			static size_t Compress(char *p_pBuffer, size_t p_nBufferSize, char *p_pOutputBuffer)
 			{
-				return LZ4_compress(p_pBuffer, p_pOutputBuffer, p_nBufferSize);
+				return (size_t)LZ4_compress(p_pBuffer, p_pOutputBuffer, p_nBufferSize);
 			}
 
 			static size_t Decompress(char *p_pBuffer, size_t p_nOriginalSize, char *p_pOutputBuffer)
 			{				
-				return LZ4_uncompress(p_pBuffer, p_pOutputBuffer, p_nOriginalSize);
+				return (size_t)LZ4_uncompress(p_pBuffer, p_pOutputBuffer, p_nOriginalSize);
 			}
 		};
 	}
